+++ src/mcp_anywhere/base.py
@@ -1,0 +1,9 @@
+"""Base database model for MCP Anywhere."""
+
+from sqlalchemy.orm import DeclarativeBase
+
+
+class Base(DeclarativeBase):
+    """Base class for all database models"""
+
+    pass
