+++ src/mcp_anywhere/core/__init__.py
@@ -1,5 +1,5 @@
 """Core MCP Anywhere functionality"""
 
-from .mcp_manager import MCPManager, store_server_tools
+from .mcp_manager import MCPManager
 
-__all__ = ["MCPManager", "store_server_tools"]
+__all__ = ["MCPManager"]
