--- tests/test_auth_routes.py
@@ -1,305 +1,0 @@
-import pytest
-import pytest_asyncio
-import tempfile
-import os
-from unittest.mock import Mock, patch
-from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine, async_sessionmaker
-from starlette.applications import Starlette
-from starlette.middleware.sessions import SessionMiddleware
-from starlette.testclient import TestClient
-from starlette.routing import Route
-from starlette.responses import HTMLResponse
-
-from mcp_anywhere.database import Base
-from mcp_anywhere.auth.models import User, OAuth2Client
-from mcp_anywhere.auth.routes import auth_routes
-
-
-@pytest_asyncio.fixture
-async def auth_db_session():
-    """Database session for auth routes tests."""
-    # Create temporary file for test database
-    temp_fd, temp_path = tempfile.mkstemp(suffix=".db")
-    os.close(temp_fd)
-    
-    engine = None
-    session = None
-    
-    try:
-        # Create async engine
-        engine = create_async_engine(f"sqlite+aiosqlite:///{temp_path}")
-        
-        # Create tables
-        async with engine.begin() as conn:
-            await conn.run_sync(Base.metadata.create_all)
-        
-        # Create session factory
-        async_session_factory = async_sessionmaker(
-            engine, class_=AsyncSession, expire_on_commit=False
-        )
-        
-        # Create and yield session
-        session = async_session_factory()
-        yield session
-        
-    finally:
-        # Cleanup
-        if session:
-            await session.close()
-        if engine:
-            await engine.dispose()
-        if os.path.exists(temp_path):
-            os.unlink(temp_path)
-
-
-@pytest_asyncio.fixture
-async def test_user(auth_db_session: AsyncSession):
-    """Create a test user."""
-    user = User(username="testuser")
-    user.set_password("testpassword")
-    auth_db_session.add(user)
-    await auth_db_session.commit()
-    await auth_db_session.refresh(user)
-    return user
-
-
-@pytest_asyncio.fixture
-async def oauth_client(auth_db_session: AsyncSession):
-    """Create a test OAuth2 client."""
-    client = OAuth2Client(
-        client_id="test_client_123",
-        client_secret="test_secret",
-        redirect_uri="http://localhost:3000/callback",
-        scope="read write"
-    )
-    auth_db_session.add(client)
-    await auth_db_session.commit()
-    await auth_db_session.refresh(client)
-    return client
-
-
-@pytest_asyncio.fixture
-async def test_app(auth_db_session: AsyncSession):
-    """Create test Starlette app with auth routes."""
-    app = Starlette()
-    
-    # Add session middleware
-    app.add_middleware(SessionMiddleware, secret_key="test-secret-key")
-    
-    # Mock the get_async_session dependency as context manager
-    from contextlib import asynccontextmanager
-    
-    @asynccontextmanager
-    async def mock_get_session():
-        yield auth_db_session
-    
-    # Add the mock function to app state
-    app.state.get_async_session = mock_get_session
-    
-    # Add a simple root route for redirects
-    async def home(request):
-        return HTMLResponse("Home")
-    
-    app.routes.append(Route("/", home))
-    
-    # Get auth routes
-    routes = auth_routes()
-    
-    # Add routes to app
-    for route in routes:
-        app.routes.append(route)
-    
-    return app
-
-
-@pytest.mark.asyncio
-async def test_login_get_renders_form(test_app: Starlette):
-    """Test that login GET endpoint renders the login form."""
-    with TestClient(test_app) as client:
-        response = client.get("/auth/login")
-        assert response.status_code == 200
-        assert "login" in response.text.lower()
-
-
-@pytest.mark.asyncio
-async def test_login_post_valid_credentials(test_app: Starlette, test_user: User):
-    """Test login with valid credentials."""
-    with TestClient(test_app) as client:
-        response = client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        }, follow_redirects=False)
-        
-        # Should redirect after successful login
-        assert response.status_code == 302
-        
-        # Check that user is logged in (session should contain user_id)
-        # Note: This is a basic test - in a real scenario you'd check the session
-
-
-@pytest.mark.asyncio
-async def test_login_post_invalid_credentials(test_app: Starlette, test_user: User):
-    """Test login with invalid credentials."""
-    with TestClient(test_app) as client:
-        response = client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "wrongpassword"
-        })
-        
-        # Should return to login form with error
-        assert response.status_code == 200
-        assert "invalid" in response.text.lower() or "error" in response.text.lower()
-
-
-@pytest.mark.asyncio
-async def test_login_post_nonexistent_user(test_app: Starlette):
-    """Test login with non-existent user."""
-    with TestClient(test_app) as client:
-        response = client.post("/auth/login", data={
-            "username": "nonexistent",
-            "password": "password"
-        })
-        
-        # Should return to login form with error
-        assert response.status_code == 200
-        assert "invalid" in response.text.lower() or "error" in response.text.lower()
-
-
-@pytest.mark.asyncio
-async def test_logout_clears_session(test_app: Starlette, test_user: User):
-    """Test that logout clears the user session."""
-    with TestClient(test_app) as client:
-        # First log in
-        client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        })
-        
-        # Then log out
-        response = client.post("/auth/logout", follow_redirects=False)
-        
-        # Should redirect
-        assert response.status_code == 302
-
-
-@pytest.mark.asyncio
-async def test_authorize_get_not_logged_in(test_app: Starlette, oauth_client: OAuth2Client):
-    """Test authorize endpoint when user is not logged in."""
-    with TestClient(test_app) as client:
-        response = client.get("/auth/authorize", params={
-            "client_id": oauth_client.client_id,
-            "redirect_uri": oauth_client.redirect_uri,
-            "response_type": "code",
-            "scope": "read"
-        }, follow_redirects=False)
-        
-        # Should redirect to login
-        assert response.status_code == 302
-        assert "/auth/login" in response.headers["location"]
-
-
-@pytest.mark.asyncio
-async def test_authorize_get_invalid_client(test_app: Starlette, test_user: User):
-    """Test authorize endpoint with invalid client_id."""
-    with TestClient(test_app) as client:
-        # First log in
-        client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        })
-        
-        # Try to authorize with invalid client
-        response = client.get("/auth/authorize", params={
-            "client_id": "invalid_client",
-            "redirect_uri": "http://localhost:3000/callback",
-            "response_type": "code",
-            "scope": "read"
-        })
-        
-        # Should return error
-        assert response.status_code == 400
-
-
-@pytest.mark.asyncio
-async def test_authorize_get_valid_logged_in(test_app: Starlette, test_user: User, oauth_client: OAuth2Client):
-    """Test authorize endpoint when user is logged in with valid client."""
-    with TestClient(test_app) as client:
-        # First log in
-        client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        })
-        
-        # Request authorization
-        response = client.get("/auth/authorize", params={
-            "client_id": oauth_client.client_id,
-            "redirect_uri": oauth_client.redirect_uri,
-            "response_type": "code",
-            "scope": "read"
-        })
-        
-        # Should render consent form
-        assert response.status_code == 200
-        assert "authorize" in response.text.lower() or "consent" in response.text.lower()
-
-
-@pytest.mark.asyncio
-async def test_authorize_post_user_grants(test_app: Starlette, test_user: User, oauth_client: OAuth2Client):
-    """Test authorization when user grants access."""
-    with TestClient(test_app) as client:
-        # First log in
-        client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        })
-        
-        # Skip this test as session_transaction is not available in Starlette TestClient
-        pytest.skip("Session transaction not available in Starlette TestClient")
-
-
-@pytest.mark.asyncio
-async def test_authorize_post_user_denies(test_app: Starlette, test_user: User, oauth_client: OAuth2Client):
-    """Test authorization when user denies access."""
-    with TestClient(test_app) as client:
-        # First log in
-        client.post("/auth/login", data={
-            "username": "testuser",
-            "password": "testpassword"
-        })
-        
-        # Skip this test as session_transaction is not available in Starlette TestClient
-        pytest.skip("Session transaction not available in Starlette TestClient")
-
-
-@pytest.mark.asyncio
-async def test_token_endpoint_valid_code(test_app: Starlette, test_user: User, oauth_client: OAuth2Client):
-    """Test token endpoint with valid authorization code."""
-    # This test would need to set up an authorization code first
-    # For now, we'll test the basic endpoint structure
-    with TestClient(test_app) as client:
-        response = client.post("/auth/token", data={
-            "grant_type": "authorization_code",
-            "code": "invalid_code",  # This will fail, but tests the endpoint
-            "client_id": oauth_client.client_id,
-            "client_secret": oauth_client.client_secret,
-            "redirect_uri": oauth_client.redirect_uri
-        })
-        
-        # Should return 400 for invalid code
-        assert response.status_code == 400
-        assert "error" in response.json()
-
-
-@pytest.mark.asyncio
-async def test_token_endpoint_invalid_grant_type(test_app: Starlette, oauth_client: OAuth2Client):
-    """Test token endpoint with invalid grant type."""
-    with TestClient(test_app) as client:
-        response = client.post("/auth/token", data={
-            "grant_type": "password",  # Invalid grant type
-            "client_id": oauth_client.client_id,
-            "client_secret": oauth_client.client_secret
-        })
-        
-        # Should return 400 for unsupported grant type
-        assert response.status_code == 400
-        assert "unsupported_grant_type" in response.json()["error"]
